server_py/requirements.txt
fastapi
uvicorn
python-multipart
Pillow
torch
transformers
einops
timm
torchvision
accelerate
python-dotenv
chromadb
sentence-transformers